Fixed DI so everything extends kodein instances

+++ app/src/main/java/com/toxicbakery/androidthings/mirror/module/weather/model/Main.kt
@@ -2,7 +2,7 @@ package com.toxicbakery.androidthings.mirror.module.weather.model
 
 data class Main(
         val temp: Double,
-        val pressure: Int,
+        val pressure: Double,
         val humidity: Int,
         val temp_min: Double,
         val temp_max: Double,
